Move utility function out of router.py
