+++ fia_api/router.py
@@ -9,7 +9,6 @@ from http import HTTPStatus
 from pathlib import Path
 from typing import Annotated, Any, Literal
 
-import requests
 from fastapi import APIRouter, Depends, HTTPException, UploadFile
 from fastapi.security import HTTPAuthorizationCredentials
 from sqlalchemy.dialects.postgresql import JSONB
@@ -38,7 +37,7 @@ from fia_api.core.services.job import (
     get_job_by_instrument,
     job_maker,
 )
-from fia_api.core.utility import safe_check_filepath
+from fia_api.core.utility import safe_check_filepath, get_packages
 from fia_api.scripts.acquisition import (
     get_script_by_sha,
     get_script_for_job,
@@ -48,7 +47,6 @@ from fia_api.scripts.pre_script import PreScript
 
 ROUTER = APIRouter()
 jwt_security = JWTBearer()
-PACKAGE_TOKEN = os.environ.get("PACKAGE_TOKEN", "shh")
 
 
 @ROUTER.get("/healthz", tags=["k8s"])
@@ -353,23 +351,8 @@ async def upload_file_to_instrument_folder(instrument: str, filename: str, file:
     return f"Successfully uploaded {filename}"
 
 
-def get_packages(org: str, image_name: str) -> Any:
-    """Helper function for getting package versions from GitHub."""
-    response = requests.get(
-        f"https://api.github.com/orgs/{org}/packages/container/{image_name}/versions",
-        headers={"Authorization": f"Bearer {PACKAGE_TOKEN}"},
-        timeout=10,
-    )
-    if response.status_code != HTTPStatus.OK:
-        raise HTTPException(
-            status_code=response.status_code,
-            detail=f"GitHub API request failed with status code {response.status_code}: {response.text}",
-        )
-    return response.json()
-
-
-@ROUTER.get("/jobs/runner_versions", tags=["jobs"])
-async def get_mantid_runner_versions(
+@ROUTER.get("/jobs/runners", tags=["jobs"])
+async def get_mantid_runners(
     credentials: Annotated[HTTPAuthorizationCredentials, Depends(jwt_security)],
 ) -> list[str]:
     """Return a list of Mantid versions if user is authenticated."""
